+++ release-notes.md
@@ -4,7 +4,9 @@
 * 179446 - Build: Bulk edit projects
 * 182793 - Build: Set values for calculated database fields
 * 178419 - Build: Ofsted pre-registration, actual date
+* 181757 - Build: finance plan agreed forecast date
 * 181754 - Build: readiness to open meeting (ROM) forecast date
+* 183590 - BUG: Start End Year component still refers to "start date" in error message
 
 ## 0.13.0
 * 147110 - Build: Readiness to open meeting (ROM)